added description and aligned the namings for the apis
