+++ nf_requirements/v1alpha1/interface_types.go
@@ -17,41 +17,42 @@ limitations under the License.
 package v1alpha1
 
 import (
-	"k8s.io/apimachinery/pkg/types"
+	"reflect"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
-type AttachmentType string
+// +kubebuilder:object:root=true
+type Interface struct {
+	metav1.TypeMeta   `json:",inline" yaml:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty" yaml:"metadata,omitempty"`
 
-// AttachmentTypeNone defines an untagged attachement (no VLAN)
-const AttachmentTypeNone AttachmentType = "none"
-
-// AttachmentTypeVLAN defines a tagged/vlan attachement
-const AttachmentTypeVLAN AttachmentType = "vlan"
-
-type CNIType string
-
-// CNITypeSRIOV defines the sriov cni
-const CNITypeSRIOV CNIType = "sriov"
-
-// CNITypeIPVLAN defines the ipvlan cni
-const CNITypeIPVLAN CNIType = "ipvlan"
-
-// CNITypeMACVLAN defines the macvlan cni
-const CNITypeMACVLAN CNIType = "macvlan"
-
-func (r *Interface) GetNetworkInstance() types.NamespacedName {
-	nsn := types.NamespacedName{}
-	if r.Spec.NetworkInstance != nil {
-		nsn.Name = r.Spec.NetworkInstance.Name
-		nsn.Namespace = r.Spec.NetworkInstance.Namespace
-	}
-	return nsn
+	Spec   InterfaceSpec   `json:"spec,omitempty" yaml:"spec,omitempty"`
+	Status InterfaceStatus `json:"status,omitempty" yaml:"status,omitempty"`
 }
 
-func (r *Interface) GetCNIType() CNIType {
-	return r.Spec.CNIType
+// TBD how do we distinguish the loopback from the vnic(s)
+type InterfaceSpec struct {
+	// NetworkInstance defines the networkInstance to which this interface belongs
+	// Name and optionally Namespace is used here
+	NetworkInstance *corev1.ObjectReference `json:"networkInstance" yaml:"networkInstance"`
+	// CNIType defines the cniType that is used to attach the interface to the pod
+	// +kubebuilder:validation:Enum=sriov;ipvlan;macvlan
+	CNIType CNIType `json:"cniType,omitempty" yaml:"cniType,omitempty"`
+	// AttachmentType defines if the interface is attached using a vlan or not
+	// +kubebuilder:validation:Enum=none;vlan
+	AttachmentType AttachmentType `json:"attachmentType,omitempty" yaml:"attachmentType,omitempty"`
 }
 
-func (r *Interface) GetAttachmentType() AttachmentType {
-	return r.Spec.AttachmentType
+type InterfaceStatus struct {
 }
+
+// Interface type metadata.
+var (
+	InterfaceKind             = reflect.TypeOf(Interface{}).Name()
+	InterfaceGroupKind        = schema.GroupKind{Group: Group, Kind: InterfaceKind}.String()
+	InterfaceKindAPIVersion   = InterfaceKind + "." + GroupVersion.String()
+	InterfaceGroupVersionKind = GroupVersion.WithKind(InterfaceKind)
+)
